TIP-869: Move Versioning bundle specs to Akeneo\Tool
