+++ src/Akeneo/Channel/Component/spec/Normalizer/Versioning/ChannelNormalizerSpec.php
@@ -4,10 +4,10 @@ namespace spec\Akeneo\Channel\Component\Normalizer\Versioning;
 
 use Akeneo\Tool\Bundle\MeasureBundle\Family\LengthFamilyInterface;
 use Akeneo\Tool\Bundle\MeasureBundle\Family\WeightFamilyInterface;
+use Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\TranslationNormalizer;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Normalizer\Standard;
 use Akeneo\Channel\Component\Normalizer\Versioning\ChannelNormalizer;
-use Akeneo\Tool\Bundle\VersionningBundle\Normalizer\Flat\TranslationNormalizer;
 use Akeneo\Channel\Component\Model\ChannelInterface;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
