+++ src/Akeneo/Tool/Bundle/VersioningBundle/spec/Normalizer/Flat/JobInstanceNormalizerSpec.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace spec\Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat;
+
+use Akeneo\Tool\Component\Batch\Model\JobInstance;
+use PhpSpec\ObjectBehavior;
+
+class JobInstanceNormalizerSpec extends ObjectBehavior
+{
+    function it_is_initializable()
+    {
+        $this->shouldHaveType('Akeneo\Tool\Bundle\VersioningBundle\Normalizer\Flat\JobInstanceNormalizer');
+    }
+
+    function it_is_a_normalizer()
+    {
+        $this->shouldImplement('Symfony\Component\Serializer\Normalizer\NormalizerInterface');
+    }
+
+    function it_supports_job_instance_normalization_into_flat(JobInstance $jobinstance)
+    {
+        $this->supportsNormalization($jobinstance, 'flat')->shouldBe(true);
+        $this->supportsNormalization($jobinstance, 'csv')->shouldBe(false);
+        $this->supportsNormalization($jobinstance, 'json')->shouldBe(false);
+        $this->supportsNormalization($jobinstance, 'xml')->shouldBe(false);
+    }
+
+    function it_normalizes_job_instance(JobInstance $jobinstance)
+    {
+        $jobinstance->getCode()->willReturn('product_export');
+        $jobinstance->getLabel()->willReturn('Product export');
+        $jobinstance->getConnector()->willReturn('myconnector');
+        $jobinstance->getType()->willReturn('EXPORT');
+        $jobinstance->getRawParameters()->willReturn(
+            [
+                'delimiter' => ';'
+            ]
+        );
+
+        $this->normalize($jobinstance)->shouldReturn(
+            [
+                'code'          => 'product_export',
+                'label'         => 'Product export',
+                'connector'     => 'myconnector',
+                'type'          => 'EXPORT',
+                'configuration' => '{"delimiter":";"}'
+            ]
+        );
+    }
+}
